# openapi.json
{
  "components": {
    "schemas": {
      "AssessedOut": {
        "description": "A paper, and its verdict once it has been screened.",
        "properties": {
          "paper": {
            "$ref": "#/components/schemas/PaperOut"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "paper",
          "verdict"
        ],
        "title": "AssessedOut",
        "type": "object"
      },
      "BatchOut": {
        "description": "A batch and how far it has got.\n\n`added` counts papers new to the system; `items` counts every paper the batch\nhas seen, and `progress` how many wait for each stage or ended in `done` or\n`dead`. `state` is the harvest's alone; `status` is the batch's as a whole,\n`processing` while its papers are still between stages. `attempts` and\n`last_error` describe harvest failures since the last page that succeeded.",
        "properties": {
          "added": {
            "title": "Added",
            "type": "integer"
          },
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "title": "Items",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "progress": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "enum": [
                "screen",
                "fetch",
                "judge",
                "done",
                "dead"
              ]
            },
            "title": "Progress",
            "type": "object"
          },
          "since": {
            "format": "date",
            "title": "Since",
            "type": "string"
          },
          "state": {
            "enum": [
              "queued",
              "harvesting",
              "done",
              "failed"
            ],
            "title": "State",
            "type": "string"
          },
          "status": {
            "enum": [
              "queued",
              "harvesting",
              "processing",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "until": {
            "format": "date",
            "title": "Until",
            "type": "string"
          }
        },
        "required": [
          "id",
          "category",
          "since",
          "until",
          "state",
          "pages",
          "added",
          "items",
          "progress",
          "status",
          "attempts",
          "last_error",
          "created_at",
          "finished_at"
        ],
        "title": "BatchOut",
        "type": "object"
      },
      "BatchRequest": {
        "description": "Which papers to pull in: those new or revised on or after a date, in every\ncategory the wishlist names.",
        "properties": {
          "since": {
            "format": "date",
            "title": "Since",
            "type": "string"
          }
        },
        "required": [
          "since"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
      "CriterionOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "want",
              "dealbreaker"
            ],
            "title": "Kind",
            "type": "string"
          },
          "requirement": {
            "title": "Requirement",
            "type": "string"
          },
          "weight": {
            "title": "Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "kind",
          "requirement",
          "weight"
        ],
        "title": "CriterionOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HealthResponse": {
        "description": "What `/health` reports. `ok` is false when the database is unreachable.",
        "properties": {
          "database": {
            "title": "Database",
            "type": "boolean"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Migration the database is on, null if never migrated",
            "title": "Revision"
          }
        },
        "required": [
          "ok",
          "database"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "PaperOut": {
        "properties": {
          "abstract": {
            "title": "Abstract",
            "type": "string"
          },
          "arxiv_id": {
            "title": "Arxiv Id",
            "type": "string"
          },
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "arxiv_id",
          "title",
          "authors",
          "categories",
          "published_at",
          "abstract"
        ],
        "title": "PaperOut",
        "type": "object"
      },
      "PaperPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AssessedOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaperPage",
        "type": "object"
      },
      "ProfileOut": {
        "description": "The wishlist papers are ranked against, and the thresholds applied to it.",
        "properties": {
          "categories": {
            "items": {
              "type": "string"
            },
            "title": "Categories",
            "type": "array"
          },
          "criteria": {
            "items": {
              "$ref": "#/components/schemas/CriterionOut"
            },
            "title": "Criteria",
            "type": "array"
          },
          "dealbreaker_threshold": {
            "title": "Dealbreaker Threshold",
            "type": "number"
          },
          "merit_weight": {
            "title": "Merit Weight",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "screen_threshold": {
            "title": "Screen Threshold",
            "type": "number"
          }
        },
        "required": [
          "name",
          "categories",
          "criteria",
          "merit_weight",
          "dealbreaker_threshold",
          "screen_threshold"
        ],
        "title": "ProfileOut",
        "type": "object"
      },
      "SessionStatus": {
        "description": "Whether the caller is signed in, and whether signing in is even possible.\n\n`configured` is false when the deployment has no passphrase hash, so the\nlogin page can say that rather than blame the phrase you typed.",
        "properties": {
          "authenticated": {
            "title": "Authenticated",
            "type": "boolean"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          }
        },
        "required": [
          "authenticated",
          "configured"
        ],
        "title": "SessionStatus",
        "type": "object"
      },
      "SignIn": {
        "description": "The one credential there is.",
        "properties": {
          "passphrase": {
            "minLength": 1,
            "title": "Passphrase",
            "type": "string"
          }
        },
        "required": [
          "passphrase"
        ],
        "title": "SignIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerdictOut": {
        "description": "Where a paper stands against the wishlist.\n\n`stage` is `screen` when only the abstract was judged \u2014 `merit` is then null\nand `total` is fit alone \u2014 and `full` once the text was. The two totals are\nnot on one scale. `per_criterion` holds each research interest's position on\nits levels, from 0 to 1: an ordering, not a percentage.",
        "properties": {
          "blocked_by": {
            "items": {
              "type": "string"
            },
            "title": "Blocked By",
            "type": "array"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "fit": {
            "title": "Fit",
            "type": "number"
          },
          "merit": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Merit"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "per_criterion": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Per Criterion",
            "type": "object"
          },
          "stage": {
            "enum": [
              "screen",
              "full"
            ],
            "title": "Stage",
            "type": "string"
          },
          "total": {
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "stage",
          "eligible",
          "total",
          "merit",
          "fit",
          "blocked_by",
          "needs_review",
          "notes",
          "per_criterion"
        ],
        "title": "VerdictOut",
        "type": "object"
      }
    }
  },
  "info": {
    "summary": "Rank documents against weighted criteria",
    "title": "sift",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/auth/session": {
      "delete": {
        "operationId": "delete_session_auth_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "auth"
        ]
      },
      "get": {
        "description": "Whether the cookie in hand is good. The login page asks this on load.",
        "operationId": "read_session_auth_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Session",
        "tags": [
          "auth"
        ]
      },
      "post": {
        "operationId": "create_session_auth_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "auth"
        ]
      }
    },
    "/batches": {
      "get": {
        "description": "The most recent batches, newest first.",
        "operationId": "get_batches_batches_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchOut"
                  },
                  "title": "Response Get Batches Batches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Batches",
        "tags": [
          "batches"
        ]
      },
      "post": {
        "description": "Queue one harvest per category the wishlist names, in the wishlist's order.\n\nThe window closes today, so each batch means the same thing when it runs.",
        "operationId": "submit_batch_batches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BatchOut"
                  },
                  "title": "Response Submit Batch Batches Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/batches/{batch_id}": {
      "get": {
        "operationId": "get_one_batch_batches__batch_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "batch_id",
            "required": true,
            "schema": {
              "title": "Batch Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get One Batch",
        "tags": [
          "batches"
        ]
      }
    },
    "/health": {
      "get": {
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health",
        "tags": [
          "health"
        ]
      }
    },
    "/papers": {
      "get": {
        "description": "Papers with their verdicts, best first; `days` keeps to recent ones.",
        "operationId": "get_papers_papers_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Published within this many days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 3650,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Published within this many days",
              "title": "Days"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "rank",
              "enum": [
                "rank",
                "newest"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Papers",
        "tags": [
          "papers"
        ]
      }
    },
    "/profile": {
      "get": {
        "operationId": "get_profile_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Profile",
        "tags": [
          "profile"
        ]
      }
    }
  }
}